U8.cs, Lz77.cs: Applied patch by castorinop. (Fixes issue 14) NusClient.cs: Added fix by madri2, and custom fix to hopefully fix Demae channel issue.
Update issue 11
Status: Started
v patch readded, working on scripting not ignoring checkboxes.

Update issue 9
Status: Fixed
This should be fixed. Needs a rebuild, can't do that from where I am now.

+++ trunk/NUS Downloader/Lz77.cs
@@ -331,7 +331,7 @@ namespace libWiiSharp
         {
             int i, p, cmp;
             cmp = 1;
-            p = N + 1 + (textBuffer[r] == 0xffff ? 0 : textBuffer[r]);
+            p = N + 1 + (textBuffer[r] == 0xffff ? 0 : (int)textBuffer[r]);
             rightSon[r] = leftSon[r] = N; matchLength = 0;
 
             for (; ; )
